requirements.txt
pyautogen 
litellm